Goblin meele attack done and more modifications for main character
Created a control for the main character to check where was he last standing before falling to the sea.
Created the respawn point of the sea.
Created coin item pickup
Created Goblin meele attack enemy.
Modified the Ranged AI controller to be a child of a general AI Controller.
Now enemies can die.